feat: auto choose race
close #46

+++ typings/mouse.pyi
@@ -381,7 +381,7 @@ def wait(button: typing.Text = "left", target_types=("up", "down", "double")):
     """
     ...
 
-def wheel(delta: int = 1) -> None:
+def wheel(delta: float = 1) -> None:
     """
     Scrolls the wheel `delta` clicks. Sign indicates direction.
     """
